+++ src/Tracy/Objects/Plane.hs
@@ -7,7 +7,7 @@ import Tracy.Types
 import Tracy.Constants
 import Tracy.Util
 
-plane :: V3 Float -> V3 Float -> Material -> Object
+plane :: V3 Double -> V3 Double -> Material -> Object
 plane o n m =
     Object { _objectMaterial = m
            , _hit = hitPlane o n m
@@ -16,16 +16,16 @@ plane o n m =
            , _areaLightImpl = Nothing
            }
 
-shadowHitPlane :: V3 Float -> V3 Float -> Ray -> Maybe Float
+shadowHitPlane :: V3 Double -> V3 Double -> Ray -> Maybe Double
 shadowHitPlane = _hitPlane
 
-_hitPlane :: V3 Float -> V3 Float -> Ray -> Maybe Float
+_hitPlane :: V3 Double -> V3 Double -> Ray -> Maybe Double
 _hitPlane o n r =
     let t = ((o - (r^.origin)) `dot` n) / ((r^.direction) `dot` n)
     in if t > epsilon
        then Just t else Nothing
 
-hitPlane :: V3 Float -> V3 Float -> Material -> Ray -> Maybe (Shade, Float)
+hitPlane :: V3 Double -> V3 Double -> Material -> Ray -> Maybe (Shade, Double)
 hitPlane o n m ray =
     let mkHitPoint t = ray^.origin + (t *^ ray^.direction)
         s hp = defaultShade { _localHitPoint = hp
